% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Title}
\usage{
build(
  folder,
  src = "manuscript.docx",
  meta_csv = "metadata.csv",
  render_only = F,
  outformats = c("html", "pdf", "jats"),
  ...
)
}
\description{
Title
}
