% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Title}
\usage{
draft(folder, overwrite = "none")
}
\arguments{
\item{overwrite}{either \code{"all"}, \code{"build"}, or \code{"none"} (default). All overwrtites all files including the manuscript template, \code{"build"} just the files in the \code{build} folder.}
}
\value{
\code{TRUE} if all files were copied successfully, \code{FALSE} otherwise.
}
\description{
Title
}
