% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownify.R
\name{markdownify}
\alias{markdownify}
\title{Converts word .docx files to .Rmd files that can be rendered to PDFs using latex}
\usage{
markdownify(
  src_docx,
  working_folder = ".",
  meta_csv = NULL,
  rmd_outpath = NULL,
  reference_parsing = F,
  url_parsing = T,
  doi_parsing = T,
  guess_refnumbers = T
)
}
\arguments{
\item{reference_parsing}{Default \code{TRUE}. IF \code{FALSE}, copy references as is without automated parsing using anystyle.}
}
\description{
Converts word .docx files to .Rmd files that can be rendered to PDFs using latex
}
