% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_yaml_cmds.R
\name{parse_yaml_cmds}
\alias{parse_yaml_cmds}
\title{Parses YAML command tags in word documents}
\usage{
parse_yaml_cmds(
  c_comtext,
  convert_straight_singlequote = T,
  escape_newlines_for_rmarkdown_propagation = T
)
}
\description{
Parses YAML command tags in word documents
}
