% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfRotuce-package.R
\docType{package}
\name{pdfRotuce-package}
\alias{pdfRotuce}
\alias{pdfRotuce-package}
\title{pdfRotuce: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
