% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabul.R
\name{rabulify}
\alias{rabulify}
\title{Create latex code for tables}
\usage{
rabulify(
  d,
  linesep = "\\newline",
  mode = "twocolumn",
  caption = NULL,
  footnote = NULL,
  label = NULL,
  long = F,
  xltabular = T,
  colwidths = NULL,
  colaligns = NULL,
  gridmode = "academicgrid",
  number_label = "",
  compat_cell_md_parsing = F
)
}
\description{
Create latex code for tables
}
