% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawn_cmd.R
\name{spawn_cmd}
\alias{spawn_cmd}
\title{spawn_cmd}
\usage{
spawn_cmd(args)
}
\arguments{
\item{args}{Character vector, first item the main command, subsequent items are arguments}
}
\value{
Entire console output of command
}
\description{
Runs a command while continously polling for the console output and forwarding it to the R console (processx based wrapper for R system/system2)
}
