% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_anystyle.R
\name{uninstall_anystyle}
\alias{uninstall_anystyle}
\title{Install the anystyle and anystyle-cli ruby gems}
\usage{
uninstall_anystyle()
}
\description{
Install the anystyle and anystyle-cli ruby gems
}
