% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Title}
\usage{
draft(folder, overwrite = F)
}
\arguments{
\item{folder}{}
}
\value{
\code{TRUE} if all files were copied successfully, \code{FALSE} otherwise.
}
\description{
Title
}
