% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownify.R
\name{markdownify}
\alias{markdownify}
\title{Converts word .docx files to .Rmd files that can be rendered to PDFs using latex}
\usage{
markdownify(
  src_docx,
  working_folder = ".",
  meta_csv = NULL,
  rmd_outpath = NULL
)
}
\arguments{
\item{meta_csv}{}
}
\description{
Converts word .docx files to .Rmd files that can be rendered to PDFs using latex
}
